refactored requests.py; changed related imports
