+++ libs/langchain/langchain/agents/agent_toolkits/nla/tool.py
@@ -5,10 +5,10 @@ from typing import Any, Optional
 
 from langchain.agents.tools import Tool
 from langchain.chains.api.openapi.chain import OpenAPIEndpointChain
-from langchain.requests import Requests
 from langchain.schema.language_model import BaseLanguageModel
 from langchain.tools.openapi.utils.api_models import APIOperation
 from langchain.tools.openapi.utils.openapi_utils import OpenAPISpec
+from langchain.utils.requests import Requests
 
 
 class NLATool(Tool):
