+++ libs/langchain/langchain/agents/agent_toolkits/nla/toolkit.py
@@ -6,11 +6,11 @@ from pydantic import Field
 
 from langchain.agents.agent_toolkits.base import BaseToolkit
 from langchain.agents.agent_toolkits.nla.tool import NLATool
-from langchain.requests import Requests
 from langchain.schema.language_model import BaseLanguageModel
 from langchain.tools.base import BaseTool
 from langchain.tools.openapi.utils.openapi_utils import OpenAPISpec
 from langchain.tools.plugin import AIPlugin
+from langchain.utils.requests import Requests
 
 
 class NLAToolkit(BaseToolkit):
