+++ libs/langchain/langchain/chains/api/openapi/chain.py
@@ -12,9 +12,9 @@ from langchain.chains.api.openapi.requests_chain import APIRequesterChain
 from langchain.chains.api.openapi.response_chain import APIResponderChain
 from langchain.chains.base import Chain
 from langchain.chains.llm import LLMChain
-from langchain.requests import Requests
 from langchain.schema.language_model import BaseLanguageModel
 from langchain.tools.openapi.utils.api_models import APIOperation
+from langchain.utils.requests import Requests
 
 
 class _ParamMapping(NamedTuple):
