+++ libs/langchain/tests/unit_tests/chains/test_api.py
@@ -8,7 +8,7 @@ import pytest
 from langchain import LLMChain
 from langchain.chains.api.base import APIChain
 from langchain.chains.api.prompt import API_RESPONSE_PROMPT, API_URL_PROMPT
-from langchain.requests import TextRequestsWrapper
+from langchain.utils.requests import TextRequestsWrapper
 from tests.unit_tests.llms.fake_llm import FakeLLM
